+++ src/Resources/contao/dca/tl_fernschach_spieler.php
@@ -182,7 +182,7 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 	'palettes' => array
 	(
 		'__selector__'                => array('death', 'fgm_title', 'sim_title', 'fim_title', 'ccm_title', 'lgm_title', 'cce_title', 'lim_title', 'gm_title', 'im_title', 'wgm_title', 'fm_title', 'wim_title', 'cm_title', 'wfm_title', 'wcm_title', 'honor_25', 'honor_40', 'honor_50', 'honor_60', 'honor_70', 'honor_president', 'honor_member', 'sepaBeitrag', 'sepaNenngeld'),
-		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,streichung,memberships,verein;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},klassenberechtigung,turnierAnmeldungenBewerbungen;{iccf_legend:hide},fgm_title,sim_title,fim_title,ccm_title,lgm_title,cce_title,lim_title,titelinfo;{fide_legend:hide},gm_title,im_title,wgm_title,fm_title,wim_title,cm_title,wfm_title,wcm_title;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},checkBeitrag;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info;{publish_legend},published,fertig'
+		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,streichung,memberships,verein;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},klassenberechtigung,turnierAnmeldungenBewerbungen;{iccf_legend:hide},fgm_title,sim_title,fim_title,ccm_title,lgm_title,cce_title,lim_title,titelinfo;{fide_legend:hide},gm_title,im_title,wgm_title,fm_title,wim_title,cm_title,wfm_title,wcm_title;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},checkBeitrag;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info;{publish_legend},published'
 	),
 
 	// Subpalettes
@@ -1940,20 +1940,20 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 			),
 			'sql'                     => "char(1) NOT NULL default ''"
 		),
-		'fertig' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['fertig'],
-			'exclude'                 => true,
-			'filter'                  => true,
-			'default'                 => '',
-			'inputType'               => 'checkbox',
-			'eval'                    => array
-			(
-				'mandatory'           => false,
-				'tl_class'            => 'w50',
-			),
-			'sql'                     => "char(1) NOT NULL default ''"
-		),
+		//'fertig' => array
+		//(
+		//	'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['fertig'],
+		//	'exclude'                 => true,
+		//	'filter'                  => true,
+		//	'default'                 => '',
+		//	'inputType'               => 'checkbox',
+		//	'eval'                    => array
+		//	(
+		//		'mandatory'           => false,
+		//		'tl_class'            => 'w50',
+		//	),
+		//	'sql'                     => "char(1) NOT NULL default ''"
+		//),
 	)
 );
 
@@ -2661,7 +2661,7 @@ class tl_fernschach_spieler extends \Backend
 		{
 			while($objAnmeldungen->next())
 			{
-				$objTurnier = \Schachbulle\ContaoFernschachbundle\Classes\Helper::getTurnierdatensatz($objAnmeldungen->pid);
+				$objTurnier = \Schachbulle\ContaoFernschachBundle\Classes\Helper::getTurnierdatensatz($objAnmeldungen->pid);
 				$records[] = array
 				(
 					'typ'        => 'Anmeldung',
@@ -2677,7 +2677,7 @@ class tl_fernschach_spieler extends \Backend
 		{
 			while($objBewerbungen->next())
 			{
-				$objTurnier = \Schachbulle\ContaoFernschachbundle\Classes\Helper::getTurnierdatensatz($objAnmeldungen->pid);
+				$objTurnier = \Schachbulle\ContaoFernschachBundle\Classes\Helper::getTurnierdatensatz($objAnmeldungen->pid);
 				$records[] = array
 				(
 					'typ'        => 'Bewerbung',